Integration with cloudiator 2.0

+++ cp_generator/src/main/java/eu/paasage/upperware/profiler/generator/properties/GeneratorProperties.java
@@ -47,6 +47,6 @@ public class GeneratorProperties {
         @NotBlank
         private String apiKey;
 
-        private long httpReadTimeout = 50;
+        private int httpReadTimeout = 50;
     }
 }
